titanic_llm_app/requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Google Gemini AI
google-generativeai>=0.3.0

# Environment Management
python-dotenv>=1.0.0

# Optional: Enhanced functionality
numpy>=1.24.0
plotly>=5.15.0  # For interactive plots (optional)
scikit-learn>=1.3.0  # For advanced analytics (optional)